+++ CHANGELOG.md
@@ -2,9 +2,15 @@
 
 ## Unreleased
 
-Allow all-uppercase function exports (fixes #11). This only works when using direct export.
+### Add `allowConstantExport` option (fixes #8)
 
-So this pattern doesn't warn anymore:
+This option allow tp don't warn when a constant (string, number, boolean, templateLiteral) is exported aside one or more components.
+
+This should be enabled if the fast refresh implementation correctly handles this case (HMR when the constant doesn't change, propagate update to importers when the constant changes.). Vite supports it, PR welcome if you notice other integrations works well.
+
+### Allow all-uppercase function exports (fixes #11)
+
+This only works when using direct export. So this pattern doesn't warn anymore:
 
 ```jsx
 export const CMS = () => <></>;